- More refactorings - Pulled up FieldLocation - Pulled up FieldLocation tests
